refactor(chat): move transient tool-approval state to a non-persisted store
The in-flight per-tool-call approval map (pending/requestApproval/
resolveApproval/clearPendingForThread) lived in the persisted useToolApproval
store, so every approval add/remove flushed the tool-approval key to disk
(including one-time 'allow once' approvals, which persist nothing). The
resolve callbacks are non-serializable and were only excluded via partialize.

Split it into a new non-persisted useToolApprovalRequests store; useToolApproval
now holds only durable settings (approvedTools, allowAllMCPPermissions) and the
legacy modal state. requestApproval/resolveApproval read the settings store for
allow-all / already-approved / allow-always. Dropped the unused isApprovalPending.

Consumers (tool.tsx, MessageItem.tsx, $threadId.tsx) repointed; tests moved to
useToolApprovalRequests.test.ts.
